Summary: create default layout.
