+++ src/App.js
@@ -1,4 +1,4 @@
-import { DefaultLayout } from "components/Layout";
+import { DefaultLayout } from "components/Layouts";
 import { Fragment } from "react";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import { privateRoutes, publicRoutes } from "routes";
