+++ src/components/GlobalStyles/GlobalStyles.scss
@@ -2,10 +2,14 @@
 @import url('https://fonts.googleapis.com/css2?family=IBM+Plex+Sans:wght@200;300;400;500;600;700&display=swap');
 
 :root {
-  --primary: black
+  --primary: black;
+  --default-layout-header: 60px;
+  
 }
 
 * {
+  margin: 0;
+  padding: 0;
   box-sizing: border-box;
 }
