--- src/components/Layout/components/Header/index.js
@@ -1,5 +1,0 @@
-function Header() {
-  return <h2>Header</h2>;
-}
-
-export default Header;
