--- src/components/Layout/components/SideBar/index.js
@@ -1,5 +1,0 @@
-function SideBar() {
-  return <h2>SideBar</h2>;
-}
-
-export default SideBar;
