+++ src/components/Layouts/DefaultLayout/DefaultLayout.module.scss
@@ -1,0 +1,17 @@
+.wrapper {
+  display: flex;
+  width: 100%;
+
+  .container {
+      width: 100%;
+      min-height: 1000px;
+      display: flex;
+      flex-direction: column;
+  
+      .content {
+        width: 100%;
+      }
+    }
+}
+
+
