+++ src/components/Layouts/DefaultLayout/index.js
@@ -1,0 +1,22 @@
+import Footer from "../components/Footer";
+import Header from "../components/Header";
+import SideBar from "../components/SideBar";
+
+import classNames from "classnames/bind";
+import styles from "./DefaultLayout.module.scss";
+
+const cx = classNames.bind(styles);
+
+function DefaultLayout({ children }) {
+  return (
+    <div className={cx("wrapper")}>
+      <SideBar />
+      <div className={cx("container")}>
+        <Header />
+        <div className={cx("content")}>{children}</div>
+      </div>
+    </div>
+  );
+}
+
+export default DefaultLayout;
