+++ src/components/Layouts/components/Header/Header.module.scss
@@ -1,0 +1,13 @@
+.wrapper {
+  height: 60px;
+  box-shadow: 0px 1px 1px rgb(0 0 0 / 12%);
+  display: flex;
+  justify-content: center;
+  width: 100%;
+}
+
+.inner {
+  height: 100%;
+  width: 1150px;
+  background-color: #ccc;
+}
