+++ src/components/Layouts/components/Header/index.js
@@ -1,0 +1,16 @@
+import classNames from "classnames/bind";
+import styles from "./Header.module.scss";
+
+const cx = classNames.bind(styles);
+
+function Header() {
+  return (
+    <header className={cx("wrapper")}>
+      <div className={cx("inner")}>
+        <h1>Header</h1>
+      </div>
+    </header>
+  );
+}
+
+export default Header;
