+++ src/components/Layouts/components/SideBar/SideBar.module.scss
@@ -1,0 +1,5 @@
+.wrapper{
+  width: 356px;
+  min-height: 1000px;
+  background-color: #ccc;
+}
