+++ src/components/Layouts/components/SideBar/index.js
@@ -1,0 +1,14 @@
+import styles from "./SideBar.module.scss";
+import classNames from "classnames/bind";
+
+const cx = classNames.bind(styles);
+
+function SideBar() {
+  return (
+    <aside className={cx("wrapper")}>
+      <h2>Side bar</h2>
+    </aside>
+  );
+}
+
+export default SideBar;
