+++ src/pages/Login/Login.module.scss
@@ -1,0 +1,6 @@
+.wrapper {
+  width: 100%;
+  display: flex;
+  align-items: center;
+  text-align: center;
+}
